+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fix timing issue that could result in multiple Logtalk terminals being created at startup
 * Fix bug where closing a non-Logtalk terminal would mark the Logtalk terminal as closed
 * Fix the delete action for the dead code scanner to correctly check for predicate arity
+* Fix predicate argument parsing bugs with commas and parentheses in quoted atoms
 
 ## [0.63.0]